chore: (Day) Move VariableDescriptor to GameObjects
