+++ Intersect (Core)/GameObjects/Switches and Variables/ServerVariableBase.cs
@@ -4,7 +4,6 @@ using System.Linq;
 
 using Intersect.Enums;
 using Intersect.GameObjects.Switches_and_Variables;
-using Intersect.Models;
 
 using Newtonsoft.Json;
