+++ Intersect (Core)/GameObjects/UserVariableBase.cs
@@ -3,7 +3,6 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using Intersect.Enums;
 using Intersect.GameObjects.Switches_and_Variables;
-using Intersect.Models;
 
 using Newtonsoft.Json;
